static cast in for range

+++ src/ray/core_worker/lib/java/jni_utils.h
@@ -289,7 +289,7 @@ inline void JavaLongArrayToNativeLongVector(JNIEnv *env, jlongArray long_array,
                                             std::vector<long> *native_vector) {
   jlong *long_array_ptr = env->GetLongArrayElements(long_array, nullptr);
   jsize vec_size = env->GetArrayLength(long_array);
-  for (size_t i = 0; i < vec_size; ++i) {
+  for (int i = 0; i < vec_size; ++i) {
     native_vector->push_back(static_cast<long>(long_array_ptr[i]));
   }
   env->ReleaseLongArrayElements(long_array, long_array_ptr, 0);
@@ -300,7 +300,7 @@ inline void JavaDoubleArrayToNativeDoubleVector(JNIEnv *env, jdoubleArray double
                                                 std::vector<double> *native_vector) {
   jdouble *double_array_ptr = env->GetDoubleArrayElements(double_array, nullptr);
   jsize vec_size = env->GetArrayLength(double_array);
-  for (size_t i = 0; i < vec_size; ++i) {
+  for (int i = 0; i < vec_size; ++i) {
     native_vector->push_back(static_cast<double>(double_array_ptr[i]));
   }
   env->ReleaseDoubleArrayElements(double_array, double_array_ptr, 0);
